Calypso examples - added usecase calypso5."

+++ java/component/keyple-smartcardio/src/main/java/org/eclipse/keyple/smartcardio/Card.h
@@ -81,7 +81,7 @@ private:
     /**
      * State of this card connection
      */
-    long unsigned int state;
+    DWORD state;
 
     /**
      * Thread holding exclusive access to the card, or null
